src/requirements.txt
colorama==0.4.4
flasgger==0.9.7.1
flask==2.3.3
flask_caching==2.0.2
gunicorn
markdown-it-py==3.0.0
psutil==5.9.5
PyYAML==5.4.1
pyzipper==0.3.6
requests==2.31.0
stdlib_list==0.9.0
tabulate==0.9.0
